pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sysml-windseeker"
version = "0.2.0"
description = "SysML v2 dependency analysis, notebook generation, execution, and view extraction tool"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

authors = [
  { name = "Windseeker Contributors" }
]

dependencies = [
  "networkx>=3.2",
  "matplotlib>=3.8",
  "scipy>=1.11",
  "nbformat>=5.9",
  "nbclient>=0.9",
  "jupyter>=1.0",
  "cairosvg>=2.7",
  "pillow>=10.0",
  "typer>=0.12",
  "rich>=13.7",
  "lxml>=5.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "pytest>=8",
  "python-semantic-release>=10.0.0",
  "pre-commit>=3.7",
  "pytest-cov>=5.0",
  "coverage[toml]>=7.6",
  "interrogate>=1.7",
]

[tool.coverage.run]
branch = true
source = ["windseeker"]

[tool.coverage.report]
show_missing = true
skip_empty = true
omit = [
  "tests/*",
]

[tool.coverage.xml]
output = "coverage.xml"

[project.scripts]
windseeker = "windseeker.cli:main"

# -------------------------
# setuptools package config
# -------------------------
[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["windseeker*"]

# -------------------------
# ruff (format + lint)
# -------------------------
[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"

# -------------------------
# Python Semantic Release
# -------------------------
[tool.semantic_release]
allow_zero_version = true
version_toml = ["pyproject.toml:project.version"]
changelog_file = "CHANGELOG.md"
assets = ["pyproject.toml", "CHANGELOG.md"]
commit_parser = "emoji"

[tool.semantic_release.branches.main]
match = "main"
prerelease = false

[tool.semantic_release.commit_parser_options]
major_tags = [":boom:"]
minor_tags = [":sparkles:"]
patch_tags = [":bug:", ":ambulance:"]
other_allowed_tags = [
  ":memo:",
  ":white_check_mark:",
  ":art:",
  ":wrench:",
  ":package:"
]
